pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "QSnake", email = "jester36883688@gmail.com"},
]
dependencies = [
    "msal>=1.23.0",
    "commitizen>=3.5.3",
    "minio>=7.1.15",
    "fastapi>=0.100.0",
    "mysql-connector-python>=8.1.0",
    "pyyaml>=6.0.1",
    "uvicorn>=0.23.1",
    "pytest>=7.4.0",
    "pytest-mock>=3.11.1",
]
requires-python = ">=3.11"
license = {text = "MIT"}
